requirement.txt
pygame>=2.5
python-chess>=1.999